requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
Werkzeug==3.0.1
requests==2.31.0
azure-identity==1.15.0
azure-ai-projects==1.0.0 